[OSFUSE-799] Camel Jasypt Encryption support in Spring Boot

+++ components-starter/camel-salesforce-starter/src/main/java/org/apache/camel/component/salesforce/springboot/SalesforceComponentConfiguration.java
@@ -467,7 +467,7 @@ public class SalesforceComponentConfiguration {
          * 
          * @param defaultReplayId
          */
-        private Integer defaultReplayId;
+        private Long defaultReplayId;
         /**
          * Replay IDs to start from per channel name.
          */
@@ -739,11 +739,11 @@ public class SalesforceComponentConfiguration {
             this.maxBackoff = maxBackoff;
         }
 
-        public Integer getDefaultReplayId() {
+        public Long getDefaultReplayId() {
             return defaultReplayId;
         }
 
-        public void setDefaultReplayId(Integer defaultReplayId) {
+        public void setDefaultReplayId(Long defaultReplayId) {
             this.defaultReplayId = defaultReplayId;
         }
